+++ invenio_app_rdm/records_ui/views/records.py
@@ -237,10 +237,10 @@ def record_detail(
         include_deleted=include_deleted,
         is_draft=is_draft,
         community=resolved_community,
-        external_resources=get_external_resources(record_ui),
+        external_resources=get_external_resources(record),
         user_avatar=avatar,
-        record_owner_username=(
-            record_owner.get("username")
+        record_owner_id=(
+            record_owner.get("id")
         ),  # record created with system_identity have not owners e.g demo
     )