+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/user_dashboard/uploads_items/MobileUploadsItem.js
@@ -7,7 +7,6 @@
 import { i18next } from "@translations/invenio_app_rdm/i18next";
 import React from "react";
 import PropTypes from "prop-types";
-import _truncate from "lodash/truncate";
 import _get from "lodash/get";
 import { Dropdown, Icon, Item, Label } from "semantic-ui-react";
 import { SearchItemCreators } from "../../utils";
@@ -77,10 +76,8 @@ export const MobileUploadsItem = ({
             <SearchItemCreators creators={creators} />
           </div>
         </Item.Meta>
-        <Item.Description>
-          {_truncate(descriptionStripped, {
-            length: 100,
-          })}
+        <Item.Description className="truncate-lines-2">
+          {descriptionStripped}
         </Item.Description>
         <Item.Extra>
           <Item.Extra>
@@ -127,14 +124,14 @@ export const MobileUploadsItem = ({
             <Dropdown.Menu>
               <Dropdown.Item
                 onClick={() => editRecord()}
-                labelPosition="left"
+                labelposition="left"
                 icon="edit"
                 content={i18next.t("Edit")}
               />
 
               {isPublished && (
                 <Dropdown.Item
-                  labelPosition="left"
+                  labelposition="left"
                   href={viewLink}
                   icon="eye"
                   content={i18next.t("View")}