+++ invenio_app_rdm/upgrade_scripts/migrate_11_0_to_12_0.py
@@ -12,13 +12,41 @@ Disclaimer: This script is intended to be executed *only once*, namely when
 upgrading from InvenioRDM 11.0 to 12.0!
 If this script is executed at any other time, probably the best case scenario
 is that nothing happens!
+
+
+This script has been tested with following data:
+
+- user
+  - demo records of v11
+  - demo communities of v11
+  - cli created user of v11
+  - ui created community (com_a)
+  - ui created community private (com_b)
+  - ui created record (rec_a.v1)
+  - ui created v2 of record (rec_a.v1)
+  - ui created record (rec_b.v1) added to community (com_a)
+  - ui created draft (dra_a)
+  - ui created draft (dra_b) added to community (com_b)
+  - repository with records without managed doi
+  - repository with records with managed doi
+  - repository with records with managed doi without parent doi after migration
+  - base vocabularies (no customized) usable after migration
+  - record (rec_a.v1, rec_a.v2) findable after migration and rebuild of index
+
+- administration
+  - user panel list of users visible
+  - drafts visible
+  - records visible
 """
+import sys
 
 from click import secho
 from flask import current_app
+from invenio_access.permissions import system_identity
 from invenio_communities.communities.records.api import Community
 from invenio_communities.communities.records.systemfields.access import ReviewPolicyEnum
 from invenio_db import db
+from invenio_rdm_records.fixtures import PrioritizedVocabulariesFixtures
 from invenio_rdm_records.proxies import current_rdm_records
 from invenio_rdm_records.records.api import RDMDraft, RDMRecord
 
@@ -31,15 +59,18 @@ def execute_upgrade():
     """
 
     def migrate_review_policy(community_record):
-        review_policy = community_record["access"].get(
+        if community_record.is_deleted:
+            return
+
+        community_record["access"].setdefault(
             "review_policy", ReviewPolicyEnum.CLOSED.value
         )
-        community_record["access"]["review_policy"] = review_policy
 
     def update_parent(record):
         """Update parent schema and parent communities for older records."""
         new_parent_schema = "local://records/parent-v3.0.0.json"
         record.parent["$schema"] = new_parent_schema
+
         if (
             isinstance(record.parent["access"]["owned_by"], list)
             and len(record.parent["access"]["owned_by"]) > 0
@@ -51,7 +82,13 @@ def execute_upgrade():
         if "pids" not in record.parent:
             record.parent["pids"] = {}
 
-            if current_app.config["DATACITE_ENABLED"]:
+            if (
+                current_app.config["DATACITE_ENABLED"]
+                and "doi" in current_app.config["RDM_PARENT_PERSISTENT_IDENTIFIERS"]
+                and current_app.config["RDM_PARENT_PERSISTENT_IDENTIFIERS"]["doi"][
+                    "is_enabled"
+                ]
+            ):
                 pids = current_rdm_records.records_service.pids.parent_pid_manager.create_all(
                     record.parent, pids={}, schemes={"doi"}
                 )
@@ -59,6 +96,17 @@ def execute_upgrade():
                     record.parent, pids
                 )
                 record.parent["pids"] = pids
+                # Have to commit here otherwise register_or_update won't get
+                # the above data
+                record.parent.commit()
+
+                if isinstance(record, RDMRecord):
+                    current_rdm_records.records_service.pids.register_or_update(
+                        id_=record["id"],
+                        identity=system_identity,
+                        scheme="doi",
+                        parent=True,
+                    )
 
     def update_record(record):
         # skipping deleted records because can't be committed
@@ -79,7 +127,6 @@ def execute_upgrade():
 
             update_parent(record)
 
-            record.parent.commit()
             record.commit()
 
             secho(f"> Updated parent: {record.parent.pid.pid_value}", fg="green")
@@ -91,6 +138,10 @@ def execute_upgrade():
             return error
 
     secho("Starting data migration...", fg="green")
+
+    # upgrading vocabularies
+    pvf = PrioritizedVocabulariesFixtures(system_identity)
+    pvf.load()
 
     # Migrating communities
     communities = Community.model_cls.query.all()
@@ -146,7 +197,7 @@ def execute_upgrade():
         )
         secho(msg, fg="yellow", err=True)
 
-    # TODO create parent DOIs? Or warn that new records will have one?
+        sys.exit(1)
 
 
 # if the script is executed on its own, perform the upgrade